requirements.txt
# LLM clients
anthropic>=0.40.0
ollama>=0.4.0

# Vector DB & embeddings
chromadb>=0.5.0
sentence-transformers>=3.0.0

# PDF processing
pymupdf>=1.24.0

# Data & validation
pydantic>=2.0.0
python-dotenv>=1.0.0

# CLI
rich>=13.0.0
typer>=0.12.0
